Pickle test data with protocol 2 for py2 compatibility
The actual cloudpickle module *does* use protocol=2 by default, but when I
generated test data for the test case, I had erroneously used protocol=3 by
hand. This resulted in passing tests on python3 but failing test on
python2. This fixes the test failure and also brings the test in line with
actual behavior.

Since cloudpickle uses protocol=2, the pickle will be deserializable on the
Jupyter side no matter which combination of python/pyspark versions is in
use.