Clean up session when session.start() throws exception
**Description**
When session start() throws exception, session delete() is invoked when
session status is not in NOT_STARTED_SESSION_STATUS to clean up resource.
This makes it safe to add the session to session manager only when session
is started properly.

This change is to address the edge case which could cause leaking session.
From client side we could see timeout issue but actually the session is
already created on Livy side.

**Testing Done**
Added unit test cases.
Also tested in a notebook.